moved output channel
