+++ clients/cobol-lsp-vscode-extension/src/services/Settings.ts
@@ -46,7 +46,7 @@ import {
 import { SupportedVariables } from "./util/FSUtils";
 import { decodeUnknown, DecodingError } from "./util/decoder";
 import * as t from "io-ts";
-import { getOutputChannel } from "../extension";
+import { getOutputChannel } from "./util/OutputChannel";
 
 const NONE: string = "NONE";
 const MAX_VM_COUNT = 50000;
