+++ clients/cobol-lsp-vscode-extension/src/services/copybook/CopybookMessageHandler.ts
@@ -14,7 +14,7 @@
 
 import * as vscode from "vscode";
 import { loadProcessorGroupCopybooksLibs } from "../ProcessorGroups";
-import { getOutputChannel } from "../../extension";
+import { getOutputChannel } from "../util/OutputChannel";
 
 export async function readFileContent(
   fileUri: string,
