+++ clients/cobol-lsp-vscode-extension/src/services/copybook/downloader/CopybookDownloaderForE4E.ts
@@ -25,7 +25,7 @@ import { CopybookURI } from "../CopybookURI";
 import { E4E_FOLDER, USE_MAP } from "../../../constants";
 import { asPartialProfile, hasMember, Utils } from "../../util/Utils";
 import { SettingsService } from "../../Settings";
-import { getOutputChannel } from "../../../extension";
+import { getOutputChannel } from "../../util/OutputChannel";
 
 const defaultConfigs: ExternalConfigurationOptions = {
   compiler: "IGYCRCTL",
