+++ clients/cobol-lsp-vscode-extension/src/services/copybookLibs/EndevorMemberLib.ts
@@ -21,7 +21,7 @@ import {
 } from "../ProcessorGroupsLoader";
 import { externalApis } from "../ExternalAPIsService";
 import { EndevorLib } from "./EndevorLib";
-import { getOutputChannel } from "../../extension";
+import { getOutputChannel } from "../util/OutputChannel";
 
 export class EndevorMemberLib extends EndevorLib implements CopybookLib {
   constructor(private config: EndevorDatasetConfigModel) {
