+++ clients/cobol-lsp-vscode-extension/src/services/util/OutputChannel.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2025 Broadcom.
+ * The term "Broadcom" refers to Broadcom Inc. and/or its subsidiaries.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Broadcom, Inc. - initial API and implementation
+ */
+
+import * as vscode from "vscode";
+
+let outputChannel: vscode.LogOutputChannel | undefined;
+
+export function getOutputChannel() {
+  if (!outputChannel) {
+    outputChannel = vscode.window.createOutputChannel(
+      "COBOL Language Support",
+      { log: true },
+    );
+  }
+  return outputChannel;
+}
